requirements.txt
google-cloud-aiplatform>=1.40.0
google-cloud-run
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
vertexai>=1.0.0
google-genai
